Refactor: rationalise BloodParticle-related constants; move BloodParticle to `src` module; add type hints
